feat: 接通 fillDensity 参数并修复 ZScript 配置截断 - 为 ZScript 和 Python 两个入口统一接入 fillDensity 参数链路 - ZScript 版新增 Fill Density 滑杆，并将其写入 zmeshmend_config.txt - Python 版新增 fillDensity 配置、事件保存和 UI 滑杆，并在调用核心 EXE 时传入 --fill-density - C++ 核心新增 fillDensity 读取与解析逻辑，支持从配置文件和命令行接收该参数 - 补洞调用改为使用 CGAL::parameters::density_control_factor(opt_fill_density) ，日志输出当前密度值 - 修复 ZScript 版配置写入过长导致文件被 NUL 截断的问题：将 SaveConfig 改为紧凑格式，并把 fillDensity 提前写入 - 对齐 Python 版与 ZScript 版的配置顺序、Settings 面板顺序和默认值 - 同步示例配置文件，新增 fillDensity=1.0 ，并将 relaxFactor 默认值调整为 0.3 - 优化 run_build.ps1 ，从开发机硬编码路径改为自动探测 cmake / vcpkg 的通用构建脚本 - 修正部分 GoZ 底层实现的类型安全问题，并清理 GoZ_Binary.h 中的编码异常文本

+++ ZMeshMendData/GoZ_Utils.cpp
@@ -166,8 +166,8 @@ bool
 GoZ_Utils::writeGoZBloc(FILE* pFile, int itemTag, int itemsCount, void *pItems, float modifier)
 {
   int headerSize = sizeof(GoZ_Header);
-  int dataSize = 0;
-  int padSize = 0;
+  size_t dataSize = 0;
+  size_t padSize = 0;
   switch (itemTag)
   {
   case GoZ_TAG_END_OF_FILE:
@@ -226,13 +226,13 @@ GoZ_Utils::writeGoZBloc(FILE* pFile, int itemTag, int itemsCount, void *pItems, 
   header.tag = itemTag;
   header.modifier = modifier;
   header.iCount = itemsCount;
-  header.size = headerSize + dataSize + padSize;
+  header.size = static_cast<unsigned int>(headerSize + dataSize + padSize);
   if (fwrite(&header, 1, sizeof(GoZ_Header), pFile) != sizeof(GoZ_Header))
     return false;
 
   if (dataSize && (fwrite(pItems, 1, dataSize, pFile)!=dataSize))
     return false;
-  if (padSize && fseek(pFile, ftell(pFile)+padSize, SEEK_SET))
+  if (padSize && fseek(pFile, ftell(pFile) + static_cast<long>(padSize), SEEK_SET))
     return false;
   return true;
 }
@@ -382,13 +382,13 @@ GoZ_Utils::readStringPref(const char* pFileName, const char* pPrefName, char* pO
     return false;
 
   // Removes start/end commas if any.
-  int start=0, end=strlen(str);
+  size_t start = 0, end = strlen(str);
   if ((end>=2) && (str[0]==34) && (str[end-1]==34))
     ++start, --end;
 
   // Truncs the string if needed.
-  if ((end-start) >= pOutPrefValueSize)
-    end = start+pOutPrefValueSize-1;
+  if ((end-start) >= static_cast<size_t>(pOutPrefValueSize))
+    end = start + static_cast<size_t>(pOutPrefValueSize) - 1;
 
   // Copy the "cleaned" string.
   str[end] = 0;
@@ -512,7 +512,7 @@ GoZ_Utils::findRunningAppProcessID(char* pAppPath, GoZ_ProcessID* pOutProcessID)
     return false;
 
   // Replaces all backslashes by slashes in the full application path.
-  unsigned int i, n;
+  size_t i, n;
   for (i=0, n=strlen(pAppPath); i<n; ++i) if (pAppPath[i] == '\\')  pAppPath[i] = '/';
 
   // Loops on every running process, and tests if it matches the same application path.
